requirements.txt
pillow==6.1.0
opencv-python==4.1.0.25
keras-vggface>=0.6
mtcnn>=0.0.9
numpy==1.17.0
scipy==1.3.1
matplotlib==3.1.1
gunicorn==19.9.0
Flask==1.0.3
flask-restful==0.3.7
Werkzeug==0.14.0